macro-nutrient/requirements.txt
Flask>=2.0.0
Flask-JWT-Extended>=4.0.0
firebase-admin
google-cloud-firestore
gunicorn
tensorflow>=2.8,<2.15
protobuf<=3.20.3
Pillow
